+++ packages/spacecat-shared-data-access/src/models/base/schema.js
@@ -315,7 +315,7 @@ class Schema {
    * @throws {SchemaError} - Throws an error if the entity is not a BaseModel or BaseCollection.
    * @return {Object[]}
    */
-  toAccessorConfigs(entity, log = console) {
+  toAccessorConfigs(entity) {
     if (!(entity instanceof BaseModel) && !(entity instanceof BaseCollection)) {
       throw new SchemaError(this, 'Entity must extend BaseModel or BaseCollection');
     }
@@ -323,7 +323,7 @@ class Schema {
     const indexAccessors = this.getIndexAccessors();
     const accessorConfigs = [];
 
-    indexAccessors.forEach(({ indexName, keySets }) => {
+    indexAccessors.forEach(({ keySets }) => {
       // generate a method for each prefix of the keySets array
       // for example, if keySets = ['opportunityId', 'status'], we create:
       //   allByOpportunityId(...)
@@ -345,8 +345,6 @@ class Schema {
           name: keyNamesToMethodName(subset, 'findBy'),
           requiredKeys: subset,
         });
-
-        log.debug(`Created accessors for index [${indexName}] with keys [${subset.join(', ')}]`);
       });
     });