+++ packages/spacecat-shared-data-access/test/unit/models/audit/audit.model.test.js
@@ -186,11 +186,13 @@ describe('AuditModel', () => {
       PAID_TRAFFIC_ANALYSIS_WEEKLY: 'paid-traffic-analysis-weekly',
       PAID_TRAFFIC_ANALYSIS_MONTHLY: 'paid-traffic-analysis-monthly',
       READABILITY: 'readability',
+      PRERENDER: 'prerender',
+      PRODUCT_METATAGS: 'product-metatags',
     };
 
     it('should have all audit types present in AUDIT_TYPES', () => {
       expect(auditTypes).to.eql(expectedAuditTypes);
-      expect(Object.keys(auditTypes)).to.have.lengthOf(35);
+      expect(Object.keys(auditTypes)).to.have.lengthOf(37);
     });
 
     it('should not have unexpected audit types in AUDIT_TYPES', () => {