+++ packages/spacecat-shared-tier-client/CHANGELOG.md
@@ -1,3 +1,31 @@
+# [@adobe/spacecat-shared-tier-client-v1.1.2](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-tier-client-v1.1.1...@adobe/spacecat-shared-tier-client-v1.1.2) (2025-09-24)
+
+
+### Bug Fixes
+
+* move logs to debug ([#982](https://github.com/adobe/spacecat-shared/issues/982)) ([7a78c6e](https://github.com/adobe/spacecat-shared/commit/7a78c6ee011bbc8adf162644153170d3ebf29c83))
+
+# [@adobe/spacecat-shared-tier-client-v1.1.1](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-tier-client-v1.1.0...@adobe/spacecat-shared-tier-client-v1.1.1) (2025-09-22)
+
+
+### Bug Fixes
+
+* **deps:** update external major (major) ([#974](https://github.com/adobe/spacecat-shared/issues/974)) ([994e488](https://github.com/adobe/spacecat-shared/commit/994e488e04fa44585b8d41a6daa9c7f86f1cee37))
+
+# [@adobe/spacecat-shared-tier-client-v1.1.0](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-tier-client-v1.0.1...@adobe/spacecat-shared-tier-client-v1.1.0) (2025-09-22)
+
+
+### Features
+
+* entitlement with out site and upgrade ([#971](https://github.com/adobe/spacecat-shared/issues/971)) ([1f042df](https://github.com/adobe/spacecat-shared/commit/1f042df62439383ff4a7cea6b0eb75649439e72a))
+
+# [@adobe/spacecat-shared-tier-client-v1.0.1](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-tier-client-v1.0.0...@adobe/spacecat-shared-tier-client-v1.0.1) (2025-09-18)
+
+
+### Bug Fixes
+
+* fix in create org entity ([#965](https://github.com/adobe/spacecat-shared/issues/965)) ([406cbf3](https://github.com/adobe/spacecat-shared/commit/406cbf3cd214f3ec3b332ca48d8a743c123a2ddd))
+
 # @adobe/spacecat-shared-tier-client-v1.0.0 (2025-09-16)